+++ src/components/Categories.jsx
@@ -4,30 +4,7 @@ import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 // import styles from "../styles/Categories.module.css";
 
-const categoriesPlaceholder = [
-  "electronics",
-  "jewelery",
-  "men_clothing",
-  "women_clothing",
-];
-
 const Categories = ({ categories }) => {
-  // const [categories, setCategories] = useState(null);
-  // const [error, setError] = useState(null);
-  useEffect(() => {
-    // fetch("https://fakestoreapi.com/products/categories")
-    //   .then((response) => {
-    //     // How to mock response status codes >= 400?
-    //     if (response.status >= 400) throw new Error("ERROR");
-    //     return response.json();
-    //   })
-    //   .then((data) => {
-    //     setCategories(data);
-    //   })
-    //   .catch((error) => setError(error.message));
-  }, []);
-
-  // if (error) return <article>{error}</article>;
   return (
     <article>
       {categories ? (
@@ -54,13 +31,3 @@ Categories.propTypes = {
 };
 
 export default Categories;
-
-/*
-(
-        <ul>
-          {categories.map((category) => (
-            <li key={category}>{category}</li>
-          ))}
-        </ul>
-      )
-*/