+++ src/components/Shop.jsx
@@ -2,6 +2,8 @@ import { useEffect, useState } from "react";
 import { Outlet, useParams } from "react-router-dom";
 import Categories from "./Categories";
 import Loading from "./Loading";
+import parseProducts from "../utilities/parseProducts";
+import parseCategory from "../utilities/parseCategory";
 import styles from "../styles/Shop.module.css";
 
 const Shop = () => {
@@ -17,9 +19,10 @@ const Shop = () => {
     //  https://fakestoreapi.com/products
     // Need to send data into Outlet and/or Categories component(s)
     Promise.all([
-      fetch("https://fakestoreapi.com/products/categories"),
-      fetch("https://fakestoreapi.com/products"),
-      // fetch(""),
+      // fetch("https://fakestoreapi.com/products/categories"),
+      // fetch("https://fakestoreapi.com/products"),
+      fetch(""),
+      fetch(""),
     ])
       .then(([resCategories, resProducts]) => {
         // How to mock response status codes >= 400?
@@ -29,16 +32,19 @@ const Shop = () => {
         // return Promise.all([resCategories.json()]);
       })
       .then(([dataCategories, dataProducts]) => {
-        // Need to create an array of objects?
-        // [{ to: men's-clothing, text: men's clothing}]
+        // How to use mockReturnValues here?
+        // const parsedProducts = parseProducts(dataProducts);
         setCategories(dataCategories);
+        // setProducts(parsedProducts);
         setProducts(dataProducts);
       })
       .catch((error) => setError(error.message));
   }, []);
 
+  // const productKey = category ? parseCategory(category) : "all";
+  // const activeProduct = products[productKey];
+
   if (error) return <section>{error}</section>;
-  console.log(category);
   return (
     <section id="shop">
       Shop section