+++ src/utilities/parseProducts.js
@@ -1,3 +1,43 @@
+import parseCategory from "./parseCategory";
+
+/* Need to create an object with categories as keys and map respective products?
+ * products = {
+ *   electronics: [ ...objects of category "electronics" ],
+ *   jewelry: [ ...objects of category "jewelry" ],
+ *   mensClothing: [ ...objects of category "men's clothing" ],
+ *   womensClothing: [ ...objects of category "women's clothing" ],
+ * }
+ */
+const parseProducts = (data) => {
+  // console.log(data);
+  const products = data.reduce(
+    (accumulator, currentProduct) => {
+      const key = parseCategory(currentProduct.category);
+      // if (accumulator[key]) {
+      //   return { ...accumulator, [key]: [...accumulator[key], currentProduct] };
+      // }
+
+      // return { ...accumulator, [key]: [currentProduct] };
+      return {
+        ...accumulator,
+        [key]: [...(accumulator[key] ?? []), currentProduct],
+      };
+    },
+    { all: [...data] }
+  );
+
+  return products;
+};
+
+export default parseProducts;
+
+/*
+
+const regExp = /(')|(\s+)([^\s])/g;
+const replacer = (match, p1, p2, p3, offSet, string) => {
+  return p3 ? p3.toUpperCase() : '';
+};
+
 const products = [
   {
     id: 1,
@@ -267,11 +307,26 @@ const products = [
   },
 ];
 
-const categories = [
-  "electronics",
-  "jewelery",
-  "men's clothing",
-  "women's clothing",
-];
+const parseCategory = (category) => {
+  return category.replace(regExp, replacer);
+};
 
-export { categories, products };
+const createDataProducts = (data) => {
+  const products = data.reduce((accumulator, currentProduct) => {
+    const key = parseCategory(currentProduct.category);
+    // if (accumulator[key]) {
+    //   return { ...accumulator, [key]: [...accumulator[key], currentProduct] };
+    // }
+
+    // return { ...accumulator, [key]: [currentProduct] };
+    return {
+      ...accumulator,
+      [key]: [...(accumulator[key] ?? []), currentProduct],
+    };
+  }, {});
+
+  return products;
+};
+
+const foo = createDataProducts(products);
+*/